# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/work-experience": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "work-experience"
                ],
                "summary": "Insert Work Experience",
                "parameters": [
                    {
                        "description": "Body Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.InsertWorkExperienceReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Personal Information Details",
                        "schema": {
                            "$ref": "#/definitions/dto.BaseResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorSchema"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorSchema"
                        }
                    }
                }
            }
        },
        "/api/work-experience/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "work-experience"
                ],
                "summary": "Update Work Experience",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Work Experience ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Body Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateWorkExperienceReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Work Experience resp",
                        "schema": {
                            "$ref": "#/definitions/dto.BaseResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorSchema"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.BaseResp": {
            "type": "object",
            "properties": {
                "error_schema": {
                    "$ref": "#/definitions/dto.ErrorSchema"
                },
                "output_schema": {
                    "description": "OutputSchema fleksibel (bisa berisi apa saja)"
                }
            }
        },
...
        "dto.ErrorMessage": {
            "type": "object",
            "properties": {
                "en": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "dto.ErrorSchema": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "string"
                },
                "error_message": {
                    "$ref": "#/definitions/dto.ErrorMessage"
                }
            }
        },
...
        "dto.InsertWorkExperienceReq": {
            "type": "object",
            "properties": {
                "company_name": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "job_description": {
                    "type": "string"
                },
                "job_title": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateWorkExperienceReq": {
            "type": "object",
            "properties": {
                "company_name": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "job_description": {
                    "type": "string"
                },
                "job_title": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}